added start of session encryption
