+++ parllama/chat_session.py
@@ -2,6 +2,7 @@
 
 from __future__ import annotations
 
+import base64
 import os
 import uuid
 from collections.abc import Iterator
@@ -38,12 +39,25 @@ from parllama.settings_manager import settings
 class ChatSession(ChatMessageContainer):
     """Chat session class"""
 
-    _subs: set[MessagePump]
     name_generated: bool
+    """Set to True if the session name has been generated by LLM"""
+    _subs: set[MessagePump]
     _abort: bool
+    """Set to True if the session should be aborted"""
     _generating: bool
+    """Set to True if the session is currently generating"""
+    _batching: bool
+    """Set to True if the session is currently being batched"""
     _stream_stats: Optional[TokenStats]
+    """Stream statistics"""
     _llm_config: LlmConfig
+    """LLM configuration"""
+    _key_secure: Optional[str]
+    """Encrypted current session password"""
+    _key: Optional[bytes]
+    """Decryption key derived from the password."""
+    _salt: Optional[bytes]
+    """Used with password to derive the en/decryption key."""
 
     # pylint: disable=too-many-arguments
     def __init__(
@@ -58,6 +72,9 @@ class ChatSession(ChatMessageContainer):
         """Initialize the chat session"""
         super().__init__(id=id, name=name, messages=messages, last_updated=last_updated)
         self._batching = True
+        self._salt = None
+        self._key = None
+        self._key_secure = None
         self.name_generated = False
         self._abort = False
         self._generating = False
@@ -291,6 +308,10 @@ class ChatSession(ChatMessageContainer):
         return json.dumps(
             {
                 "id": self.id,
+                "_salt": (
+                    base64.b64encode(self._salt).decode("utf-8") if self._salt else None
+                ),
+                "__key__": self._key_secure,
                 "name": self.name,
                 "name_generated": self.name_generated,
                 "last_updated": self.last_updated.isoformat(),
@@ -305,6 +326,7 @@ class ChatSession(ChatMessageContainer):
     def from_json(json_data: str, load_messages: bool = False) -> ChatSession:
         """Convert JSON to chat session"""
         data: dict = json.loads(json_data)
+
         messages = data["messages"]
         for m in messages:
             # convert old format
@@ -332,6 +354,14 @@ class ChatSession(ChatMessageContainer):
             llm_config=LlmConfig.from_json(lc),
         )
         session.name_generated = True
+
+        if salt := data.get("__salt__"):
+            session._salt = base64.b64decode(salt)  # pylint: disable=protected-access
+        else:
+            session._salt = None  # pylint: disable=protected-access
+
+        session._key_secure = data.get("__key__")  # pylint: disable=protected-access
+
         return session
 
     @staticmethod
